Move InfoFragment inside TransactionInfo package
